# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://labnote.test/api/v1",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/auth/signup": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Devuelve un usuario creado",
                "description": "Devuelve un usuario creado",
                "operationId": "App\\Http\\Controllers\\AuthController::signup",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "names": {
                                        "description": "",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "lastname": {
                                        "description": "",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "mother_lastname": {
                                        "description": "",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "rut": {
                                        "description": "",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "email": {
                                        "description": "",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "password": {
                                        "description": "",
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Devuelve un token jwt",
                "description": "Devuelve un token de autenticación",
                "operationId": "App\\Http\\Controllers\\AuthController::login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "rut": {
                                        "description": "",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "password": {
                                        "description": "",
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        []
    ]
}